feat(main): 메인 로직 작성
main함수에 각 로또 횟수별로 당첨 결과를 출력하는 기능을 작성

+++ src/main/kotlin/lotto/Lotto.kt
@@ -74,7 +74,7 @@ fun getBonusNumber():Int {
         throw IllegalArgumentException(ErrorMessage.notInRange)
     }
 
-    if (bonusNumber in pickedNumbers) {
+    if (bonusNumber.toInt() in pickedNumbers) {
         throw IllegalArgumentException(ErrorMessage.isDuplex)
     }
     return bonusNumber.toInt()
